moved execute_with_multiprocessing_lock to eolearn.core
