+++ core/eolearn/core/eoexecution.py
@@ -20,6 +20,7 @@ import threading
 import traceback
 import concurrent.futures
 import datetime as dt
+import multiprocessing
 
 from tqdm.auto import tqdm
 
@@ -28,6 +29,8 @@ from .eoworkflow import EOWorkflow
 from .utilities import LogFileFilter
 
 LOGGER = logging.getLogger(__name__)
+
+MULTIPROCESSING_LOCK = multiprocessing.Manager().Lock()
 
 
 class EOExecutor:
@@ -269,3 +272,19 @@ class EOExecutor:
                                'reports')
 
         return EOExecutorVisualization(self).make_report()
+
+
+def execute_with_multiprocessing_lock(execution_function, *args, **kwargs):
+    """ A helper utility function that executes a given function with multiprocessing lock if the process is being
+    executed in a multi-processing mode
+
+    :param execution_function: A function
+    :param args: Function's positional arguments
+    :param kwargs: Function's keyword arguments
+    :return: Function's results
+    """
+    if multiprocessing.current_process().name == 'MainProcess':
+        return execution_function(*args, **kwargs)
+
+    with MULTIPROCESSING_LOCK:
+        return execution_function(*args, **kwargs)
